consensus:engine:backend: call get priority transactors from contract on Finalize
